+++ docs/change_log.rst
@@ -48,6 +48,10 @@ The following are only available within Stem's `git repository
   * Support for protocol descriptor fields (:spec:`eb4fb3c`)
   * Shared randomness properties weren't being read in votes (:trac:`21102`)
 
+ * **Utilities**
+
+  * Added :class:`~stem.util.test_tools.TimedTestRunner` and :func:`~stem.util.test_tools.test_runtimes`
+
 .. _version_1.5:
 
 Version 1.5 (November 20th, 2016)